backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic-settings
python-dotenv
langchain
langchain-community
langchain-openai
langchain-google-genai
python-multipart
requests
gunicorn
affinda
firecrawl-py
deepgram-sdk
elevenlabs
hume
pinecone-client
resend
python-jose[cryptography]
